Fix error uploading offset datetime (#263)
* Fix error uploading offset datetime

### Description

Without this change Metabase will throw the following error when uploading a CSV with a datetime with an offset value:

```
class java.time.OffsetDateTime` cannot be cast to class `java.lang.String`
```

This is because we were missing a case in the insert method, and it was falling through the the catch all case. The first fix is to add an explicit case for this type to fix the JDBC integration.

The second fix is to update the config for how we map CSV columns to database types in particular. Since Clickhouse does not preserve the input offset, we made the product choice to keep these values as their raw input string. By returning `nil` for this type we can trigger this behavior.

* Fix fall through case