Some more test fixes:
1. use ddl.i/format-name.
One of our tests verified we can connect. This failed on a database
called `test-data` which was turned into `test_data` in clickhouse.

2. Comment out datetime64, gmt0 mapping to column without
timezone. This has a timezone and that is necessary in
`metabase.query-processor-test.timezones-test/filter-test`. This sets
expectations about which timezone to use. It should use timezones if
specified and fall back to report timezone. But we ignore timezones if
the column doesn't have a timezone.

3. Turn off tests which expect parameter-count feature

4. Set some different results for a test where we get
offset-date-times back